refactor create new test button into its own component
