+++ packages/reporter/src/header/CreateNewTestButton.tsx
@@ -1,0 +1,21 @@
+import Button from '@cypress-design/react-button'
+import { IconActionAddMedium } from '@cypress-design/react-icon'
+import React, { MouseEvent, useCallback } from 'react'
+import events from '../lib/events'
+import cs from 'classnames'
+
+export const CreateNewTestButton = ({ suiteId }: { suiteId: string }) => {
+  const _launchStudio = useCallback((e: MouseEvent) => {
+    e.preventDefault()
+    e.stopPropagation()
+
+    events.emit('studio:init:suite', suiteId)
+  }, [events, suiteId])
+
+  return (
+    <Button data-cy='create-new-test-button' size='20' onClick={_launchStudio} variant='outline-dark' className={cs('launch-studio-button')} >
+      <IconActionAddMedium strokeColor='gray-500' />
+      New Test
+    </Button>
+  )
+}
