+++ packages/reporter/src/runnables/runnable-and-suite.tsx
@@ -1,7 +1,7 @@
 import cs from 'classnames'
 import _ from 'lodash'
 import { observer } from 'mobx-react'
-import React, { MouseEvent, useCallback, useMemo } from 'react'
+import React, { useCallback, useMemo } from 'react'
 
 import appState from '../lib/app-state'
 import events, { Events } from '../lib/events'
@@ -11,9 +11,9 @@ import Collapsible, { CollapsibleHeaderComponentProps } from '../collapsible/col
 import type SuiteModel from './suite-model'
 import type TestModel from '../test/test-model'
 
-import { IconActionAddMedium, IconChevronDownMedium, IconChevronRightMedium, IconObjectStackFailed, IconObjectStackPassed, IconObjectStackQueued, IconObjectStackRunning, IconObjectStackSkipped, WindiColor } from '@cypress-design/react-icon'
-import Button from '@cypress-design/react-button'
+import { IconChevronDownMedium, IconChevronRightMedium, IconObjectStackFailed, IconObjectStackPassed, IconObjectStackQueued, IconObjectStackRunning, IconObjectStackSkipped, WindiColor } from '@cypress-design/react-icon'
 import { RunnableArray } from './runnables-store'
+import { CreateNewTestButton } from '../header/CreateNewTestButton'
 
 // should only show connection dots if the current runnable is a test and the next runnable is a test and is not the last runnable
 export const shouldShowConnectionDots = (runnables: RunnableArray, runnable: SuiteModel | TestModel, runnableIndex: number) => {
@@ -35,13 +35,6 @@ const headerIconDefaultProps = {
 }
 
 const Suite: React.FC<SuiteProps> = observer(({ eventManager = events, model, studioEnabled, canSaveStudioLogs, spec }: SuiteProps) => {
-  const _launchStudio = useCallback((e: MouseEvent) => {
-    e.preventDefault()
-    e.stopPropagation()
-
-    eventManager.emit('studio:init:suite', model.id)
-  }, [eventManager, model.id])
-
   const headerIconStyle = {
     marginTop: '1px',
   }
@@ -85,16 +78,13 @@ const Suite: React.FC<SuiteProps> = observer(({ eventManager = events, model, st
         <span className='runnable-title'>{model.title}</span>
         {(studioEnabled && !appState.studioActive && spec?.relative !== '__all') && (
           <>
-            <Button data-cy='create-new-test-button' size='20' onClick={_launchStudio} variant='outline-dark' className={cs('launch-studio-button')} >
-              <IconActionAddMedium strokeColor='gray-500' />
-              New Test
-            </Button>
+            <CreateNewTestButton suiteId={model.id} />
             <span className='button-hover-shadow' />
           </>
         )}
       </>
     )
-  }, [getHeaderIcon, model.title, studioEnabled, appState.studioActive, _launchStudio])
+  }, [getHeaderIcon, model.title, studioEnabled, appState.studioActive])
 
   const runnablesList = useMemo(() => (
     <ul className='runnables'>
